# api-specification.yaml
openapi: 3.0.3
info:
  title: Catholic Information Platform API
  description: |
    REST API untuk mengakses konten Katolik termasuk Alkitab, Katekismus Gereja Katolik, artikel, dokumen Gereja, santo-santa, dan kalender liturgi.
    
    ## Otentikasi
    API ini menggunakan **API Key** untuk otentikasi. Sertakan API key Anda di header `X-API-Key`.
    
    Untuk mendapatkan API key:
    1. Register/login ke platform
    2. Kunjungi halaman API Keys di dashboard
    3. Generate API key baru
    
    ## Rate Limiting
    - Default: 1000 requests/hour per API key
    - Custom limits tersedia untuk enterprise
    
    ## Response Format
    Semua response menggunakan format JSON:
    ```json
    {
      "success": true,
      "data": {...},
      "meta": {...},
      "error": null
    }
    ```
    
    ## Endpoints
    Base URL Production: `https://api.catholic-platform.or.id/v1`
    Base URL Development: `http://localhost:3000/api/v1`
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@catholic-platform.or.id
    url: https://catholic-platform.or.id/api/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.catholic-platform.or.id/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key untuk autentikasi
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token untuk admin authentication

# ============================================================================
# REUSABLE SCHEMAS
# ============================================================================
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
...
    BibleVerse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        reference:
          type: string
          example: "Kej 1:1"
        bookName:
          type: string
          example: "Kejadian"
        chapterNumber:
          type: integer
          example: 1
        verseNumber:
          type: integer
          example: 1
        text:
          type: string
          example: "Pada mulanya Allah menciptakan langit dan bumi."
        footnotes:
          type: string
          nullable: true
        crossReferences:
          type: array
          items:
            type: string
          example: ["Yoh 1:1", "Ibr 11:3"]
    CatechismParagraph:
      type: object
      properties:
        id:
          type: integer
          example: 1
        paragraphNumber:
          type: integer
          example: 1
        title:
          type: string
          nullable: true
          example: "Kehidupan manusia - mengenal dan mengasihi Allah"
        content:
          type: string
          example: "Allah, yang tak terbatas sempurna dan berbahagia dalam diri-Nya sendiri..."
        bibleReferences:
          type: array
          items:
            type: string
          example: ["Kej 1:27", "Yoh 17:3"]
        crossReferences:
          type: array
          items:
            type: integer
          example: [27, 356, 1701]
        inBrief:
          type: boolean
          example: false
        section:
          type: string
          example: "Bab Satu: Manusia Mampu Mengenal Allah"
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Arti Ekaristi dalam Kehidupan Katolik"
        slug:
          type: string
          example: "arti-ekaristi-dalam-kehidupan-katolik"
        excerpt:
          type: string
          example: "Ekaristi adalah sumber dan puncak kehidupan Kristiani..."
        body:
          type: string
          example: "# Pengantar\n\nEkaristi merupakan..."
        featuredImage:
          type: string
          format: uri
          nullable: true
        author:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
              example: "Pastor Antonius"
        category:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
              example: "Sakramen"
            slug:
              type: string
              example: "sakramen"
        tags:
          type: array
          items:
            type: string
          example: ["Ekaristi", "Sakramen", "Liturgi"]
        status:
          type: string
          enum: [draft, published, archived]
          example: "published"
        viewCount:
          type: integer
          example: 150
        publishedAt:
          type: string
          format: date-time
          example: "2025-01-01T10:00:00Z"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Saint:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Santo Fransiskus Xaverius"
        title:
          type: string
          example: "Patron Misionaris"
        feastDay:
          type: object
          properties:
            month:
              type: integer
              example: 12
            date:
              type: integer
              example: 3
        feastRank:
          type: string
          enum: [solemnity, feast, memorial, optional_memorial]
          example: "memorial"
        liturgicalColor:
          type: string
          enum: [white, red, green, purple, rose, black]
          example: "white"
        birthYear:
          type: integer
          nullable: true
          example: 1506
        deathYear:
          type: integer
          nullable: true
          example: 1552
        biography:
          type: string
          example: "Santo Fransiskus Xaverius lahir di Spanyol..."
        patronOf:
          type: array
          items:
            type: string
          example: ["Misionaris", "Pelaut", "India"]
        symbols:
          type: array
          items:
            type: string
          example: ["Salib", "Kapal"]
        imageUrl:
          type: string
          format: uri
          nullable: true
...
paths:
...
  /bible/search:
    get:
      tags:
        - Bible
      summary: Search Bible verses
      description: Full-text search dalam Alkitab
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
          example: "kasih Allah"
        - name: testament
          in: query
          schema:
            type: string
            enum: [old_testament, new_testament, deuterocanonical]
          description: Filter by testament
        - name: bookId
          in: query
          schema:
            type: integer
          description: Filter by book ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Results per page
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BibleVerse'
                  meta:
                    $ref: '#/components/schemas/Pagination'
      security:
        - ApiKeyAuth: []

  # ==========================================================================
  # CATECHISM
  # ==========================================================================
...
  /catechism/search:
    get:
      tags:
        - Catechism
      summary: Search Catechism content
      description: Full-text search dalam KGK
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
          example: "Ekaristi"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CatechismParagraph'
                  meta:
                    $ref: '#/components/schemas/Pagination'
      security:
        - ApiKeyAuth: []

  # ==========================================================================
  # CONTENTS (Articles, Documents, Prayers, etc.)
  # ==========================================================================
  /contents:
    get:
      tags:
        - Contents
      summary: Get all contents
      description: Mendapatkan daftar konten (artikel, dokumen, doa, dll)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [article, document, prayer, homily, qa, page]
          description: Filter by content type
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
            default: published
          description: Filter by status (admin only can see draft)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, popular]
            default: newest
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  meta:
                    $ref: '#/components/schemas/Pagination'
      security:
        - ApiKeyAuth: []
...
  /saints:
    get:
      tags:
        - Saints
      summary: Get all saints
      description: Mendapatkan daftar santo-santa
      parameters:
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Filter by feast month
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
          description: Search by saint name
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Saint'
                  meta:
                    $ref: '#/components/schemas/Pagination'
      security:
        - ApiKeyAuth: []
...
  /search:
    get:
      tags:
        - Search
      summary: Global search across all content
      description: |
        Search di semua konten: Bible, Catechism, Articles, Documents, Saints, dll.
        
        Results dikelompokkan by type dengan relevance ranking.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
          example: "Ekaristi"
        - name: types
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [bible, catechism, article, document, prayer, homily, qa, saint]
          description: Filter by content types (comma-separated)
          example: "bible,catechism,article"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                              example: "bible"
                            title:
                              type: string
                              example: "Yoh 3:16"
                            excerpt:
                              type: string
                              example: "Karena begitu besar kasih Allah akan dunia ini..."
                            url:
                              type: string
                              example: "/bible/Yoh/3#16"
                            score:
                              type: number
                              example: 0.95
                      summary:
                        type: object
                        properties:
                          bible:
                            type: integer
                            example: 15
                          catechism:
                            type: integer
                            example: 8
                          article:
                            type: integer
                            example: 5
                  meta:
                    $ref: '#/components/schemas/Pagination'
      security:
        - ApiKeyAuth: []

  # ==========================================================================
  # ADMIN ENDPOINTS (Requires JWT Authentication)
  # ==========================================================================
...